+++ tests/conftest.py
@@ -18,6 +18,9 @@ https://docs.pytest.org/en/latest/reference/fixtures.html#conftest-py-sharing-fi
 """
 
 import pytest
+from iris.cube import Cube, CubeList
+
+from CSET.operators import constraints, filters, read
 
 
 @pytest.fixture()
@@ -27,17 +30,29 @@ def tmp_working_dir(tmp_path, monkeypatch):
     return tmp_path
 
 
+# For speed only load the data from disk once, then make an in-memory copy
+# whenever it is reused.
 @pytest.fixture(scope="session")
-def cubes():
-    """Get an iris CubeList."""
-    import CSET.operators.read as read
-
+def cubes_readonly():
+    """Get an iris CubeList. It is NOT safe to modify."""
     return read.read_cubes("tests/test_data/air_temp.nc")
 
 
 @pytest.fixture(scope="session")
-def cube(cubes):
-    """Get an iris Cube."""
-    from CSET.operators import constraints, filters
+def cube_readonly(cubes_readonly):
+    """Get an iris Cube. It is NOT safe to modify."""
+    return filters.filter_cubes(
+        cubes_readonly, constraints.generate_cell_methods_constraint([])
+    )
 
-    return filters.filter_cubes(cubes, constraints.generate_cell_methods_constraint([]))
+
+@pytest.fixture()
+def cubes(cubes_readonly: CubeList) -> CubeList:
+    """Get an iris CubeList. It is safe to modify."""
+    return cubes_readonly.copy()
+
+
+@pytest.fixture()
+def cube(cube_readonly: Cube) -> Cube:
+    """Get an iris Cube. It is safe to modify."""
+    return cube_readonly.copy()