+++ tests/operators/test_misc.py
@@ -30,7 +30,6 @@ def test_noop_operator():
 
 def test_remove_attribute_cube(cube):
     """Remove attribute from a Cube."""
-    cube = cube.copy()
     assert "STASH" in cube.attributes
     cube = misc.remove_attribute(cube, "STASH")[0]
     assert "STASH" not in cube.attributes
@@ -38,7 +37,6 @@ def test_remove_attribute_cube(cube):
 
 def test_remove_attribute_cubelist(cubes):
     """Remove attribute from a CubeList."""
-    cubes = cubes.copy()
     for cube in cubes:
         assert "STASH" in cube.attributes
     cubes = misc.remove_attribute(cubes, "STASH")